Consolidate .historical.util with .util
