+++ item/historical/T012.py
@@ -1,8 +1,7 @@
 """Data cleaning code and configuration for T012."""
 import numpy as np
 
-from item.historical.util import dropna_logged
-from item.utils import convert_units
+from item.util import convert_units, dropna_logged
 
 #: iTEM data flow matching the data from this source.
 DATAFLOW = "POPULATION"
