+++ item/util.py
@@ -1,7 +1,10 @@
+import logging
 from typing import Optional, Sequence
 
 import pandas as pd
 from iam_units import registry
+
+log = logging.getLogger(__name__)
 
 
 # TODO Add an argument to control the format of the output units
@@ -40,3 +43,21 @@ def convert_units(
 
     # Assign magnitude and unit columns in output DataFrame
     return df.assign(**{cols[0]: qty.magnitude, cols[1]: f"{qty.units:~}"})
+
+
+def dropna_logged(df, column, log_columns=[]):
+    """Drop rows from `df` with NaN values in `column`.
+
+    Counts and unique values for each of `log_columns` are logged.
+    """
+    # Rows to drop
+    to_drop = df[column].isnull()
+
+    log.info(f"{to_drop.sum()} rows with NaN in {repr(column)}")
+
+    for col in log_columns if to_drop.sum() else []:
+        # Sorted unique values in column `col`
+        values = sorted(df[to_drop][col].unique())
+        log.info(f"… with {len(values)} unique values in {repr(col)}: {values}")
+
+    return df[~to_drop]
